Deleted the AutoPCH folder in .vs (precompiled headers i think), .vs folder needed to make program work
Added a normal enemy spawner, and spawning all enemy's at start.
float to int warning cleanup

+++ BuasGameSDL2/Enemy.h
@@ -6,7 +6,7 @@ class Enemy :public GameObject {
 public:
 	Enemy(GameObject gameObject);
 	~Enemy();
-	float scale;
+	int scale;
 	void Update() override;
 	void Render() override;
 	void onCollision(std::string otherTag, GameObject* other);
