+++ BuasGameSDL2/GameObject.h
@@ -10,7 +10,7 @@ class GameObject {
 	public:
 		GameObject(const std::string & textureSheet, SDL_Renderer* ren, int x, int y, bool& gameStarted);
 		GameObject(SDL_Texture * textureSheet, SDL_Renderer* ren, int x, int y, bool& gameStarted);
-		float scale = 2;
+		int scale = 2;
 	 	virtual void Update();
 		virtual void Render();
 		virtual void ProccesInput(int keyCode, bool down);
